# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Operating system remote execution interface.",
        "title": "OS Remote Executor API",
        "contact": {
            "name": "osreapi",
            "url": "https://github.com/xmapst/osreapi/issues"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://github.com/xmapst/osreapi/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/task/{task}/step/{step}": {
            "get": {
                "description": "detail task step",
                "consumes": [
                    "application/json",
                    "application/toml",
                    "application/x-yaml",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "application/x-yaml",
                    "application/toml"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "task step detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "step id",
                        "name": "step",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    }
                }
            },
            "put": {
                "description": "manager task step",
                "consumes": [
                    "application/json",
                    "application/toml",
                    "application/x-yaml",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "application/x-yaml",
                    "application/toml"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "task step manager",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "step id",
                        "name": "step",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "paused",
                            "kill",
                            "pause",
                            "resume"
                        ],
                        "type": "string",
                        "default": "paused",
                        "description": "management action",
                        "name": "action",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "1m",
                        "description": "how long to pause; if empty, manual continuation is required",
                        "name": "duration",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    }
                }
            }
        },
        "/api/v1/task/{task}/workspace": {
            "get": {
                "description": "task workspace list",
                "consumes": [
                    "application/json",
                    "application/toml",
                    "application/x-yaml",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "application/x-yaml",
                    "application/toml"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "task workspace list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dir path",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    }
                }
            },
            "post": {
                "description": "task workspace upload file or dir",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "application/x-yaml",
                    "application/toml"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "task workspace upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dir path",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "files",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "task workspace delete",
                "consumes": [
                    "application/json",
                    "application/toml",
                    "application/x-yaml",
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json",
                    "application/x-yaml",
                    "application/toml"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "task workspace delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task id",
                        "name": "task",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dir path",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/types.BaseRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.BaseRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 255
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "message"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
    }
}